requirements.txt
yt-dlp
ffmpeg
Pillow
# For packaging stuff as an exe
pyinstaller